Generate decimal values in Factory
The `ActivityFactory` was generating int values for `boost` and `total_boosted_points`. While this works, it can make testing more complex since the underlying model returns a `Decimal` field.

If the property of the factory is compared to the actual value returned by the database, the comparison will fail because the factory property lacks decimal information (forcing writing decimal formatting with integers in a testing scenario).